windows-release/ham_hat_control_center/requirements.txt
pyserial>=3.5
numpy>=1.24
sounddevice>=0.4.6
